+++ dogfooding/lib/screens/lobby_screen.dart
@@ -1,12 +1,10 @@
 // 📦 Package imports:
 import 'package:flutter/material.dart';
-import 'package:flutter_svg/flutter_svg.dart';
 import 'package:stream_video_filters/video_effects_manager.dart';
 import 'package:stream_video_flutter/stream_video_flutter.dart';
 
 import '../app/user_auth_controller.dart';
 import '../di/injector.dart';
-import '../utils/assets.dart';
 import '../widgets/lobby_device_controls.dart';
 
 class LobbyScreen extends StatefulWidget {
@@ -97,8 +95,6 @@ class _LobbyScreenState extends State<LobbyScreen> {
     final textTheme = context.streamTextTheme;
     final colorTheme = context.streamColorScheme;
     final currentUser = _userAuthController.currentUser;
-
-    final theme = StreamLobbyViewTheme.of(context);
 
     return Scaffold(
       appBar: AppBar(