+++ packages/stream_video/lib/src/logger/impl/file_logger.dart
@@ -3,7 +3,9 @@ import 'dart:io';
 
 import 'package:collection/collection.dart';
 import 'package:intl/intl.dart';
-import 'package:stream_core/stream_core.dart';
+import 'package:stream_core/stream_core.dart' show Standard;
+
+import '../logger_api.dart';
 
 const String _tag = 'SV:FileLogger';
 const int _defaultSize = 12 * 1024 * 1024;