+++ packages/stream_video/lib/src/sfu/ws/sfu_ws.dart
@@ -61,12 +61,13 @@ class SfuWebSocket {
   SfuWebSocket._({
     required this.sessionSeq,
     required this.sessionId,
-    required String url,
+    required this.url,
     Iterable<String>? protocols,
   }) {
     _logger.i(() => '<init> sessionId: $sessionId');
     _client = StreamWebSocketClient(
-      options: WebSocketOptions(url: url, protocols: protocols),
+      tag: '$_tag-$sessionSeq',
+      optionsBuilder: () => WebSocketOptions(url: url, protocols: protocols),
       messageCodec: const SfuMessageCodec(),
       pingRequestBuilder: ([_]) => SfuWsRequest(
         sfu_events.SfuRequest(
@@ -82,9 +83,9 @@ class SfuWebSocket {
 
   final int sessionSeq;
   final String sessionId;
+  final String url;
 
   late final StreamWebSocketClient _client;
-  String get url => _client.options.url;
 
   /// The underlying socket client.
   ///
@@ -216,6 +217,26 @@ class SfuWebSocket {
             url: url,
             reason: const DisconnectionReason(
               closeReason: 'Unhealthy connection',
+            ),
+          ),
+        );
+      case ConnectTimeout():
+        _events.emit(
+          SfuSocketDisconnected(
+            sessionId: sessionId,
+            url: url,
+            reason: const DisconnectionReason(
+              closeReason: 'Connection attempt timed out',
+            ),
+          ),
+        );
+      case AuthenticationFailed():
+        _events.emit(
+          SfuSocketFailed(
+            sessionId: sessionId,
+            url: url,
+            error: VideoErrors.compose(
+              source.error ?? 'SFU WS authentication failed',
             ),
           ),
         );